+++ exonum-java-binding/core/src/test/java/com/exonum/binding/core/storage/indices/BaseProofMapIndexProxyIntegrationTestable.java
@@ -16,7 +16,6 @@
 
 package com.exonum.binding.core.storage.indices;
 
-import static com.exonum.binding.core.storage.indices.CheckedMapProofMatcher.isValid;
 import static com.exonum.binding.core.storage.indices.MapEntries.putAll;
 import static com.exonum.binding.core.storage.indices.MapTestEntry.absentEntry;
 import static com.exonum.binding.core.storage.indices.MapTestEntry.presentEntry;
@@ -32,7 +31,11 @@ import static com.exonum.binding.core.storage.indices.TestStorageItems.V4;
 import static com.exonum.binding.core.storage.indices.TestStorageItems.values;
 import static com.exonum.binding.test.Bytes.bytes;
 import static com.google.common.base.Preconditions.checkArgument;
+import static java.util.Collections.emptyList;
 import static java.util.Collections.singletonList;
+import static java.util.Comparator.comparing;
+import static java.util.stream.Collectors.toList;
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.hamcrest.CoreMatchers.containsString;
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.CoreMatchers.notNullValue;
@@ -46,24 +49,26 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import com.exonum.binding.common.collect.MapEntry;
 import com.exonum.binding.common.hash.HashCode;
 import com.exonum.binding.common.hash.Hashing;
-import com.exonum.binding.common.proofs.map.CheckedMapProof;
-import com.exonum.binding.common.proofs.map.UncheckedMapProof;
+import com.exonum.binding.common.serialization.Serializer;
 import com.exonum.binding.common.serialization.StandardSerializers;
 import com.exonum.binding.core.proxy.Cleaner;
 import com.exonum.binding.core.proxy.CloseFailuresException;
 import com.exonum.binding.core.storage.database.View;
+import com.exonum.core.messages.MapProofOuterClass;
+import com.exonum.core.messages.MapProofOuterClass.OptionalEntry;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Streams;
 import com.google.common.primitives.UnsignedBytes;
+import com.google.protobuf.ByteString;
+import com.google.protobuf.Empty;
 import java.util.Arrays;
-import java.util.Collections;
+import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.function.BiConsumer;
 import java.util.function.Consumer;
 import java.util.function.Function;
-import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.Test;
@@ -187,7 +192,7 @@ abstract class BaseProofMapIndexProxyIntegrationTestable
 
   @Test
   @DisabledProofTest
-  void getProof_EmptyMapDoesNotContainSingleKey() {
+  void verifyProof_EmptyMapDoesNotContainSingleKey() {
     runTestWithView(database::createSnapshot,
         (map) -> assertThat(map, provesThatAbsent(key1))
     );
@@ -195,7 +200,7 @@ abstract class BaseProofMapIndexProxyIntegrationTestable
 
   @Test
   @DisabledProofTest
-  void getProof_SingletonMapContains() {
+  void verifyProof_SingletonMapContains() {
     runTestWithView(database::createFork, (map) -> {
       HashCode key = key1;
       String value = V1;
@@ -207,7 +212,7 @@ abstract class BaseProofMapIndexProxyIntegrationTestable
 
   @Test
   @DisabledProofTest
-  void getProof_SingletonMapDoesNotContain() {
+  void verifyProof_SingletonMapDoesNotContain() {
     runTestWithView(database::createFork, (map) -> {
       map.put(key1, V1);
 
@@ -217,7 +222,7 @@ abstract class BaseProofMapIndexProxyIntegrationTestable
 
   @Test
   @DisabledProofTest
-  void getProof_MultiEntryMapContains() {
+  void verifyProof_MultiEntryMapContains() {
     runTestWithView(database::createFork, (map) -> {
       List<MapEntry<HashCode, String>> entries = createMapEntries();
       putAll(map, entries);
@@ -230,7 +235,7 @@ abstract class BaseProofMapIndexProxyIntegrationTestable
 
   @Test
   @DisabledProofTest
-  void getMultiProof_MultiEntryMapContains() {
+  void verifyMultiProof_MultiEntryMapContains() {
     runTestWithView(database::createFork, (map) -> {
       List<MapEntry<HashCode, String>> entries = createMapEntries();
       putAll(map, entries);
@@ -241,7 +246,7 @@ abstract class BaseProofMapIndexProxyIntegrationTestable
 
   @Test
   @DisabledProofTest
-  void getProof_MultiEntryMapDoesNotContain() {
+  void verifyProof_MultiEntryMapDoesNotContain() {
     runTestWithView(database::createFork, (map) -> {
       List<MapEntry<HashCode, String>> entries = createMapEntries();
       putAll(map, entries);
@@ -264,14 +269,14 @@ abstract class BaseProofMapIndexProxyIntegrationTestable
 
   @Test
   @DisabledProofTest
-  void getMultiProof_EmptyMapDoesNotContainSeveralKeys() {
+  void verifyMultiProof_EmptyMapDoesNotContainSeveralKeys() {
     runTestWithView(database::createSnapshot, (map) ->
         assertThat(map, provesThatAbsent(key1, key2)));
   }
 
   @Test
   @DisabledProofTest
-  void getMultiProof_SingletonMapDoesNotContainSeveralKeys() {
+  void verifyMultiProof_SingletonMapDoesNotContainSeveralKeys() {
     runTestWithView(database::createFork, (map) -> {
       map.put(key1, V1);
 
@@ -281,7 +286,7 @@ abstract class BaseProofMapIndexProxyIntegrationTestable
 
   @Test
   @DisabledProofTest
-  void getMultiProof_SingletonMapBothContainsAndDoesNot() {
+  void verifyMultiProof_SingletonMapBothContainsAndDoesNot() {
     runTestWithView(database::createFork, (map) -> {
       ImmutableMap<HashCode, String> source = ImmutableMap.of(
           key1, V1
@@ -295,7 +300,7 @@ abstract class BaseProofMapIndexProxyIntegrationTestable
 
   @Test
   @DisabledProofTest
-  void getMultiProof_TwoElementMapContains() {
+  void verifyMultiProof_TwoElementMapContains() {
     runTestWithView(database::createFork, (map) -> {
       ImmutableMap<HashCode, String> source = ImmutableMap.of(
           key1, V1,
@@ -367,57 +372,124 @@ abstract class BaseProofMapIndexProxyIntegrationTestable
   }
 
   @Test
-  @DisabledProofTest
+  void keysTest() {
+    runTestWithView(database::createFork, (map) -> {
+      List<MapEntry<HashCode, String>> entries = createSortedMapEntries();
+
+      putAll(map, entries);
+
+      Iterator<HashCode> keysIterator = map.keys();
+      List<HashCode> keysFromIter = ImmutableList.copyOf(keysIterator);
+      List<HashCode> keysInMap = MapEntries.extractKeys(entries);
+
+      // Keys must appear in a lexicographical order.
+      assertThat(keysFromIter, equalTo(keysInMap));
+    });
+  }
+
+  @Test
+  void valuesTest() {
+    runTestWithView(database::createFork, (map) -> {
+      List<MapEntry<HashCode, String>> entries = createSortedMapEntries();
+
+      putAll(map, entries);
+
+      Iterator<String> valuesIterator = map.values();
+      List<String> valuesFromIter = ImmutableList.copyOf(valuesIterator);
+      List<String> valuesInMap = MapEntries.extractValues(entries);
+
+      // Values must appear in a lexicographical order of keys.
+      assertThat(valuesFromIter, equalTo(valuesInMap));
+    });
+  }
+
+  @Test
+  void entriesTest() {
+    runTestWithView(database::createFork, (map) -> {
+      List<MapEntry<HashCode, String>> entries = createSortedMapEntries();
+
+      putAll(map, entries);
+
+      Iterator<MapEntry<HashCode, String>> entriesIterator = map.entries();
+      List<MapEntry> entriesFromIter = ImmutableList.copyOf(entriesIterator);
+      // Entries must appear in a lexicographical order of keys.
+      assertThat(entriesFromIter, equalTo(entries));
+    });
+  }
+
+  @Test
   void getProofFromSingleKey() {
     runTestWithView(database::createFork, (map) -> {
       map.put(key1, V1);
 
-      UncheckedMapProof proof = map.getProof(key1);
-      CheckedMapProof checkedProof = proof.check();
+      MapProof proof = map.getProof(key1);
+      MapProofOuterClass.MapProof asMessage = proof.getAsMessage();
 
-      assertThat(checkedProof, isValid(singletonList(presentEntry(key1, V1))));
+      assertThat(asMessage.getEntriesList()).containsExactly(presentOptEntry(key1, V1));
     });
   }
 
   @Test
-  @DisabledProofTest
-  void getProofFromVarargs() {
+  void getProofFromSingleMissingKey() {
+    runTestWithView(database::createFork, (map) -> {
+      map.put(key1, V1);
+
+      MapProof proof = map.getProof(key2);
+      MapProofOuterClass.MapProof asMessage = proof.getAsMessage();
+
+      assertThat(asMessage.getEntriesList()).containsExactly(absentOptEntry(key2));
+    });
+  }
+
+  @Test
+  void getProofFromVarargsPresentKeys() {
     runTestWithView(database::createFork, (map) -> {
       map.put(key1, V1);
       map.put(key2, V2);
 
-      UncheckedMapProof proof = map.getProof(key1, key2);
-      CheckedMapProof checkedProof = proof.check();
+      MapProof proof = map.getProof(key1, key2);
 
-      assertThat(
-          checkedProof, isValid(Arrays.asList(presentEntry(key1, V1), presentEntry(key2, V2))));
+      MapProofOuterClass.MapProof asMessage = proof.getAsMessage();
+      assertThat(asMessage.getEntriesList()).containsExactlyInAnyOrder(
+          presentOptEntry(key1, V1), presentOptEntry(key2, V2));
     });
   }
 
   @Test
-  @DisabledProofTest
-  void getProofFromEmptyCollection() {
+  void getProofFromVarargsPresentAndMissingKeys() {
     runTestWithView(database::createFork, (map) -> {
       map.put(key1, V1);
 
-      IllegalArgumentException thrown = assertThrows(IllegalArgumentException.class, () -> {
-        UncheckedMapProof proof = map.getProof(Collections.emptyList());
-      });
-      assertThat(thrown.getLocalizedMessage(),
-          containsString("Keys collection should not be empty"));
+      MapProof proof = map.getProof(key1, key2);
+
+      MapProofOuterClass.MapProof asMessage = proof.getAsMessage();
+      assertThat(asMessage.getEntriesList()).containsExactlyInAnyOrder(
+          presentOptEntry(key1, V1), absentOptEntry(key2));
     });
   }
 
   @Test
-  @DisabledProofTest
   void getProofFromCollection() {
     runTestWithView(database::createFork, (map) -> {
       map.put(key1, V1);
 
-      UncheckedMapProof proof = map.getProof(singletonList(key1));
-      CheckedMapProof checkedProof = proof.check();
+      MapProof proof = map.getProof(singletonList(key1));
 
-      assertThat(checkedProof, isValid(singletonList(presentEntry(key1, V1))));
+      MapProofOuterClass.MapProof asMessage = proof.getAsMessage();
+      assertThat(asMessage.getEntriesList()).containsExactly(
+          presentOptEntry(key1, V1));
+    });
+  }
+
+  @Test
+  void getProofFromEmptyCollection() {
+    runTestWithView(database::createFork, (map) -> {
+      map.put(key1, V1);
+
+      IllegalArgumentException thrown = assertThrows(IllegalArgumentException.class,
+          () -> map.getProof(emptyList()));
+      assertThat(thrown.getLocalizedMessage(),
+          containsString("Keys collection should not be empty"));
     });
   }
 
@@ -478,7 +550,17 @@ abstract class BaseProofMapIndexProxyIntegrationTestable
     index.put(key1, V1);
   }
 
-  List<MapEntry<HashCode, String>> createMapEntries() {
+  /**
+   * Creates `numOfEntries` map entries, sorted by key in lexicographical order:
+   * [(key0, V1), (key2, V2), … (key_i, Vi)].
+   */
+  private List<MapEntry<HashCode, String>> createSortedMapEntries() {
+    Stream<HashCode> sortedKeys = getTestKeys().stream()
+        .sorted(comparing(HashCode::asBytes, UnsignedBytes.lexicographicalComparator()));
+    return createMapEntries(sortedKeys);
+  }
+
+  private List<MapEntry<HashCode, String>> createMapEntries() {
     return createMapEntries(getTestKeys().stream());
   }
 
@@ -491,6 +573,25 @@ abstract class BaseProofMapIndexProxyIntegrationTestable
     Stream<String> roundRobinValues = IntStream.range(0, Integer.MAX_VALUE)
         .mapToObj(i -> values.get(i % values.size()));
     return Streams.zip(keys, roundRobinValues, MapEntry::valueOf)
-        .collect(Collectors.toList());
+        .collect(toList());
+  }
+
+  private static OptionalEntry presentOptEntry(HashCode key, String value) {
+    Serializer<String> stringSerializer = StandardSerializers.string();
+    return optEntryForKey(key)
+        .setValue(ByteString.copyFrom(stringSerializer.toBytes(value)))
+        .build();
+  }
+
+  private static OptionalEntry absentOptEntry(HashCode key) {
+    return optEntryForKey(key)
+        .setNoValue(Empty.getDefaultInstance())
+        .build();
+  }
+
+  private static OptionalEntry.Builder optEntryForKey(HashCode key) {
+    Serializer<HashCode> hashSerializer = StandardSerializers.hash();
+    return OptionalEntry.newBuilder()
+        .setKey(ByteString.copyFrom(hashSerializer.toBytes(key)));
   }
 }