+++ exonum-java-binding/testkit/src/test/java/com/exonum/binding/testkit/TestKitExtensionTest.java
@@ -32,7 +32,6 @@ import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
 import org.junit.jupiter.api.extension.RegisterExtension;
@@ -55,7 +54,6 @@ class TestKitExtensionTest {
   }
 
   @Test
-  @Disabled("Disabled until ProofMapIndexProxy 32 byte key restriction is relaxed")
   void testKitInstantiationTestCase() {
     Events testEvents = getTestCaseEvents(TestKitInstantiationTestCase.class);